🧱♻️ F07: ADMIN > Rename components associated with Admin + improved architecture
